chore(packaging): Do not install tests.*
find_packages needs to list the modules to exclude.
'docs' doesnt exist, and wouldnt be relevent unless it contained
a `__init__.py`.  OTOH, tests.* needs to be excluded to avoid
it being included in the installation's egg-info/top_level.txt
and installing a package tests conflicts with other packages
which also have broken packaging, and has a special rule in most
packaging systems to prevent it.